Add files via upload
In this file, I use Gibbs sampler to get the parameters of an logit regression. Then the logit regression is used to calculate the probability of rising oil prices in France.
The data which I used contains the historical prices of oil (Gasoil) in France and the historical prices of crude (Brent).